WIP settings API migration; pre statestore merge

+++ include/settings.h
@@ -17,6 +17,6 @@
 int start_settings_subsys();
 
 // int load_immediate_value(const char *name, void *dest, size_t len);
-ssize_t settings_load_one_init_default(const char *name, void *dest, size_t len, void *init, size_t len_init);
+uint8_t settings_load_one_init_default(const char *name, void *dest, size_t len, void *init, size_t len_init);
 
 #endif // INCLUDE_SETTINGS_H__
